requirements.txt
# InterPreT项目依赖

# 核心依赖
numpy>=1.21.0
torch>=1.9.0
transformers>=4.20.0
gym>=0.21.0

# 数据处理
pandas>=1.3.0
pyarrow>=6.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 配置和工具
pyyaml>=6.0
tqdm>=4.62.0

# 仿真环境
igibson>=2.2.0

# 测试框架
pytest>=6.2.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
